refactor(statescript): rename Look At to Set Rotation Toward and Track Target to Look At
