+++ addons/forge/editor/statescript/node_editors/LookAt3DNodeEditor.cs
@@ -12,6 +12,9 @@ namespace Gamesmiths.Forge.Godot.Editor.Statescript.NodeEditors;
 [Tool]
 internal sealed partial class LookAt3DNodeEditor : StandardNodeEditorBase
 {
+	// Input property index of the turn rate, matching LookAt3DNode.
+	private const int SpeedInputIndex = 2;
+
 	private static readonly IReadOnlyList<NodeConfigParam> _configParams =
 	[
 		new NodeConfigParam("flatten", "Ignore height", DefaultBool: true),
@@ -19,8 +22,15 @@ internal sealed partial class LookAt3DNodeEditor : StandardNodeEditorBase
 	];
 
 	public override string HandledRuntimeTypeName =>
-		"Gamesmiths.Forge.Godot.Core.Statescript.Nodes.Action.LookAt3DNode";
+		"Gamesmiths.Forge.Godot.Core.Statescript.Nodes.State.LookAt3DNode";
 
 	protected override IReadOnlyList<NodeConfigParam> ConstructorParams => _configParams;
+
+	protected override string? GetInputLabel(int inputIndex)
+	{
+		// Radians per second rather than degrees, matching Rotate To 3D: the turn is measured in the angles core's own
+		// resolvers produce, and Deg To Rad is how a degree figure gets there.
+		return inputIndex == SpeedInputIndex ? "Speed (rad/s)" : null;
+	}
 }
 #endif
