+++ addons/forge/editor/statescript/node_editors/SetRotationToward2DNodeEditor.cs
@@ -1,0 +1,25 @@
+// Copyright © Gamesmiths Guild.
+
+#if TOOLS
+using System.Collections.Generic;
+using Godot;
+
+namespace Gamesmiths.Forge.Godot.Editor.Statescript.NodeEditors;
+
+/// <summary>
+/// Node editor for Set Rotation Toward 2D.
+/// </summary>
+[Tool]
+internal sealed partial class SetRotationToward2DNodeEditor : StandardNodeEditorBase
+{
+	private static readonly IReadOnlyList<NodeConfigParam> _configParams =
+	[
+		new NodeConfigParam("nodePath", "Node", IsText: true, Placeholder: "%Turret"),
+	];
+
+	public override string HandledRuntimeTypeName =>
+		"Gamesmiths.Forge.Godot.Core.Statescript.Nodes.Action.SetRotationToward2DNode";
+
+	protected override IReadOnlyList<NodeConfigParam> ConstructorParams => _configParams;
+}
+#endif
