+++ addons/forge/editor/statescript/node_editors/SetRotationToward3DNodeEditor.cs
@@ -1,0 +1,26 @@
+// Copyright © Gamesmiths Guild.
+
+#if TOOLS
+using System.Collections.Generic;
+using Godot;
+
+namespace Gamesmiths.Forge.Godot.Editor.Statescript.NodeEditors;
+
+/// <summary>
+/// Node editor for Set Rotation Toward 3D.
+/// </summary>
+[Tool]
+internal sealed partial class SetRotationToward3DNodeEditor : StandardNodeEditorBase
+{
+	private static readonly IReadOnlyList<NodeConfigParam> _configParams =
+	[
+		new NodeConfigParam("flatten", "Ignore height", DefaultBool: true),
+		new NodeConfigParam("nodePath", "Node", IsText: true, Placeholder: "%YawPivot"),
+	];
+
+	public override string HandledRuntimeTypeName =>
+		"Gamesmiths.Forge.Godot.Core.Statescript.Nodes.Action.SetRotationToward3DNode";
+
+	protected override IReadOnlyList<NodeConfigParam> ConstructorParams => _configParams;
+}
+#endif
