--- addons/forge/editor/statescript/node_editors/TrackTarget2DNodeEditor.cs.uid
@@ -1,1 +1,0 @@
-uid://d3d7d343juu1s
