--- addons/forge/editor/statescript/node_editors/TrackTarget3DNodeEditor.cs.uid
@@ -1,1 +1,0 @@
-uid://mrpvxpmhmru3
