+++ docs/statescript/README.md
@@ -116,7 +116,7 @@ In Forge for Godot, you use the **Statescript graph editor** to create graphs vi
 4. Set the ability's `AbilityBehavior` to a `StatescriptAbilityBehavior`.
 5. Assign your `StatescriptGraph` to the behavior's `Statescript` property.
 
-The `ForgeEntity` node drives graph updates on two rails: `Abilities.UpdateAbilities(delta)` in its `_Process` method, and `Abilities.FixedUpdateAbilities(delta)` in its `_PhysicsProcess` method. If you implement `IForgeEntity` directly, you must call **both** yourself — the built-in nodes that move a body, steer an agent or query physics (Move To, Rotate To, Track Target, Nav Move To, Force Override, Overlap, Ray, Sweep, Line Of Sight) only run on the fixed rail, and a host that never calls it leaves them activated and stuck.
+The `ForgeEntity` node drives graph updates on two rails: `Abilities.UpdateAbilities(delta)` in its `_Process` method, and `Abilities.FixedUpdateAbilities(delta)` in its `_PhysicsProcess` method. If you implement `IForgeEntity` directly, you must call **both** yourself — the built-in nodes that move a body, steer an agent or query physics (Move To, Rotate To, Look At, Nav Move To, Force Override, Overlap, Ray, Sweep, Line Of Sight) only run on the fixed rail, and a host that never calls it leaves them activated and stuck.
 
 > **Threaded physics is not supported.** Overlap, Ray, Sweep and Line Of Sight run their first query during activation, so that a watch reports what is already there on the frame it begins rather than one frame later. Activation happens wherever the graph was driven from, which is usually the frame rail. Godot only guarantees `World2D`/`World3D.direct_space_state` inside `_physics_process` when `physics/2d/run_on_separate_thread` or `physics/3d/run_on_separate_thread` is on, so those settings must stay off.
